TRACKING ENTITIES IN TECHNICAL PROCEDURES - A NEW DATASET AND BASELINES
We introduce TechTrack, a new dataset for tracking entities in technical procedures. The dataset,
prepared by annotating open domain articles from WikiHow, consists of 1351 procedures, e.g., “How
to connect a printer”, identifies more than 1200 unique entities with an average of 4.7 entities per
procedure. We evaluate the performance of state-of-the-art models on the entity-tracking task and
find that they are well below the human annotation performance. We describe how TechTrack can be
used to take forward the research on understanding procedures from temporal texts.