Language for Modelling False Data Injection Attacks in Internet of Things
Internet of Things (IoT) is now omnipresent in all aspects of life and provides a large number of potentially critical services. For this, Internet of Things relies on the data collected by objects. Data integrity is therefore essential. Unfortunately, this integrity is threatened by a type of attack known as False Data Injection Attack. This consists of an attacker who injects fabricated data into a system to modify its behaviour. In this work, we dissect and present a method that uses a Domain-Specific Language (DSL) to generate altered data, allowing these attacks to be simulated and tested.