Update* This is the final train() method of your neural network
There’s a lot going on in the above code block, so here’s a line-by-line breakdown:

Line 8 picks a random instance from the dataset.

Lines 14 to 16 calculate the partial derivatives and return the derivatives for the bias and the weights. They use _compute_gradients(), which you defined earlier.

Line 18 updates the bias and the weights using _update_parameters(), which you defined in the previous code block.

Line 21 checks if the current iteration index is a multiple of 100. You do this to observe how the error changes every 100 iterations.

Line 24 starts the loop that goes through all the data instances.

Line 28 computes the prediction result.

Line 29 computes the error for every instance.

Line 31 is where you accumulate the sum of the errors using the cumulative_error variable. You do this because you want to plot a point with the error for all the data instances. Then, on line 32, you append the error to cumulative_errors, the array that stores the errors. You’ll use this array to plot the graph.

In short, you pick a random instance from the dataset, compute the gradients, and update the weights and the bias. You also compute the cumulative error every 100 iterations and save those results in an array. You’ll plot this array to visualize how the error changes during the training process.